+++ Source/AccelByteUe4Sdk/Public/Api/AccelByteUserApi.h
@@ -79,14 +79,14 @@ namespace AccelByte
 			* @param Password Password.
 			* @param OnSuccess This will be called when the operation succeeded.
 			* @param OnError This will be called when the operation failed.
-			* @param RememberMe This will use for refresh token expiration extension, default value is false. 
+			* @param bRememberMe This will use for refresh token expiration extension, default value is false. 
 			*/
 			void LoginWithUsernameV3(
 				const FString& Username,
 				const FString& Password,
 				const FVoidHandler& OnSuccess,
 				const FErrorHandler& OnError,
-				const bool RememberMe = false) const;
+				const bool bRememberMe = false) const;
 
 			/**
 			* @brief Log in with email/username account using v3 endpoint with 2FA enable
@@ -95,14 +95,14 @@ namespace AccelByte
 			* @param Password Password.
 			* @param OnSuccess This will be called when the operation succeeded.
 			* @param OnError This will be called when the operation failed.
-			* @param RememberMe This will use for refresh token expiration extension, default value is false. 
+			* @param bRememberMe This will use for refresh token expiration extension, default value is false. 
 			*/
 			void LoginWithUsernameV3(
 				const FString& Username,
 				const FString& Password,
 				const FVoidHandler& OnSuccess,
 				const FCustomErrorHandler& OnError,
-				const bool RememberMe = false) const;
+				const bool bRememberMe = false) const;
 
 			/**
 			 * @brief Log in with another platform account e.g. Steam, Google, Facebook, Twitch, etc.
@@ -125,12 +125,14 @@ namespace AccelByte
 			* @param PlatformToken Authentication code that provided by another platform.
 			* @param OnSuccess This will be called when the operation succeeded.
 			* @param OnError This will be called when the operation failed.
+			* @param bCreateHeadless If directly create new account when not linked yet, default value is true 
 			*/
 			void LoginWithOtherPlatform(
 				EAccelBytePlatformType PlatformType,
 				const FString& PlatformToken,
 				const FVoidHandler& OnSuccess,
-				const FCustomErrorHandler& OnError) const;
+				const FCustomErrorHandler& OnError,
+				bool bCreateHeadless = true) const;
 
 			/**
 			 * @brief Log in with device ID (anonymous log in).
@@ -148,10 +150,10 @@ namespace AccelByte
 			 * @param Code auth code from 3rd party authenticator or backupCode. 
 			 * @param OnSuccess This will be called when the operation succeeded.
 			 * @param OnError This will be called when the operation failed.
-			 * @param RememberDevice
+			 * @param bRememberDevice
 			 */
 			void VerifyLoginWithNewDevice2FAEnabled(const FString& MfaToken, EAccelByteLoginAuthFactorType AuthFactorType, const FString& Code,
-				const FVoidHandler& OnSuccess, const FCustomErrorHandler& OnError, bool RememberDevice = false) const;
+				const FVoidHandler& OnSuccess, const FCustomErrorHandler& OnError, bool bRememberDevice = false) const;
 
 			/**
 			 * @brief Log in from Accelbyte Launcher.
@@ -168,8 +170,26 @@ namespace AccelByte
 			 * @param OnError This will be called when the operation failed.
 			 */
 			void LoginWithRefreshToken(const FVoidHandler& OnSuccess, const FErrorHandler& OnError) const;
-			#pragma endregion /Login Methods
+
+			/**
+			 * @brief Create Headless Account And Login
+			 *
+			 * @param OnSuccess This will be called when the operation succeeded.
+			 * @param OnError This will be called when the operation failed.
+			 */
+			void CreateHeadlessAccountAndLogin(const FVoidHandler& OnSuccess, const FCustomErrorHandler& OnError) const;
 			
+			/**
+			 * @brief Authentication With Platform Link And Login
+			 *
+			 * @param Username This is username's account exist
+			 * @param Password This is password's account exist
+			 * @param OnSuccess This will be called when the operation succeeded.
+			 * @param OnError This will be called when the operation failed.
+			 */
+			void AuthenticationWithPlatformLinkAndLogin(const FString& Username, const FString& Password, const FVoidHandler& OnSuccess, const FCustomErrorHandler& OnError) const;
+
+			#pragma endregion /Login Methods
 			
 			/**
 			 * @brief General handler for LoginWith* success; mostly a multicast callback handler.
@@ -395,11 +415,24 @@ namespace AccelByte
 			 * @brief This function will search user by their Username or Display Name. The query will be used to find the user with the most approximate username or display name.
 			 *
 			 * @param Query Targeted user's Username or Display Name.
-			 * @param By Filter the responded PagedPublicUsersInfo by SearchType. Choose the SearchType.ALL if you want to be responded with all query type.
+			 * @param Offset Targeted offset query filter.
+			 * @param Limit Targeted limit query filter.
 			 * @param OnSuccess This will be called when the operation succeeded. The result is FPagedPublicUsersInfo.
 			 * @param OnError This will be called when the operation failed.
 			 */
-			void SearchUsers(const FString& Query, EAccelByteSearchType By, const THandler<FPagedPublicUsersInfo>& OnSuccess, const FErrorHandler& OnError);
+			void SearchUsers(const FString& Query, int32 Offset, int32 Limit, const THandler<FPagedPublicUsersInfo>& OnSuccess, const FErrorHandler& OnError);
+			
+			/**
+			 * @brief This function will search user by their Username or Display Name. The query will be used to find the user with the most approximate username or display name.
+			 *
+			 * @param Query Targeted user's Username or Display Name.
+			 * @param By Filter the responded PagedPublicUsersInfo by SearchType. Choose the SearchType.ALL if you want to be responded with all query type.
+			 * @param OnSuccess This will be called when the operation succeeded. The result is FPagedPublicUsersInfo.
+			 * @param OnError This will be called when the operation failed.
+			 * @param Offset Targeted offset query filter.
+			 * @param Limit Targeted limit query filter.
+			 */
+			void SearchUsers(const FString& Query, EAccelByteSearchType By, const THandler<FPagedPublicUsersInfo>& OnSuccess, const FErrorHandler& OnError, const int32& Offset = 0, const int32& Limit = 100);
 
 			/**
 			 * @brief This function will search user by userId.
@@ -484,9 +517,9 @@ namespace AccelByte
 			 * @param  LanguageCode Targeted Language Code, using ISO-639 
 			 * @param  OnSuccess This will be called when the operation succeeded. The result is FInputUserValidation.
 			 * @param  OnError This will be called when the operation failed.
-			 * @param  DefaultOnEmpty Targeted DefaultOnEmpty. 
+			 * @param  bDefaultOnEmpty Targeted DefaultOnEmpty. 
 			 */
-			void GetInputValidations(const FString& LanguageCode, THandler<FInputValidation> const& OnSuccess, FErrorHandler const& OnError, bool DefaultOnEmpty = true);
+			void GetInputValidations(const FString& LanguageCode, THandler<FInputValidation> const& OnSuccess, FErrorHandler const& OnError, bool bDefaultOnEmpty = true);
 
 			/**
 			* @brief This function will enable 2FA with backupCode.